refactor(core): extract export workflow owner
