+++ ci/proof.toml
@@ -175,6 +175,7 @@ packages = ["tokmd-core"]
 proof = [
   "cargo test -p tokmd-core lang_workflow --verbose",
   "cargo test -p tokmd-core module_workflow --verbose",
+  "cargo test -p tokmd-core export_workflow --verbose",
   "cargo test -p tokmd-core ffi",
   "cargo test -p tokmd-core --test ffi_parity_w53",
 ]
